so I'm changing a line here and another there and using github to transfer the files to the production server.

+++ functions/switch_pronunciation_in_database.php
@@ -23,32 +23,38 @@ $pronunc_variant=SQLite3::escapeString($_GET["pronunc"]);
 
 $timestamp=date('Y-m-d h:i:s', time());
 
-
-$sqlcommand ="SELECT count(*) FROM evaluations WHERE word='$word' AND listener='$listener' AND speaker='$speaker';";
-if ($db->querySingle($sqlcommand) > 0) {
-    $sqlcommand ="UPDATE evaluations SET pronunc_variant='$pronunc_variant', timestamp='$timestamp' ";
-    $sqlcommand.="WHERE word='$word' AND listener='$listener' AND speaker='$speaker';";
-
-    $sqlcommand.="DELETE FROM phone_error  ";
-    $sqlcommand.="WHERE word='$word' AND listener='$listener' AND speaker='$speaker';";
-
-}
-else {
-    $sqlcommand ="INSERT INTO evaluations (speaker, word, listener, evaluation, pronunc_variant, timestamp) ";
-    $sqlcommand.="VALUES ('$speaker', '$word', '$listener', '0', '$pronunc_variant', '$timestamp'); ";
-
-}
-
-#print $sqlcommand."<br>";
-
-
-$trying=$db->exec($sqlcommand);
-
-if ($trying)
+if ($listener == 1) {
     print 1;
-else
-    print 0;
+    }
+else  {
 
-$db->close();
+    $sqlcommand ="SELECT count(*) FROM evaluations WHERE word='$word' AND listener='$listener' AND speaker='$speaker';";
+    if ($db->querySingle($sqlcommand) > 0) {
+        $sqlcommand ="UPDATE evaluations SET pronunc_variant='$pronunc_variant', timestamp='$timestamp' ";
+        $sqlcommand.="WHERE word='$word' AND listener='$listener' AND speaker='$speaker';";
+
+        $sqlcommand.="DELETE FROM phone_error  ";
+        $sqlcommand.="WHERE word='$word' AND listener='$listener' AND speaker='$speaker';";
+
+    }
+    else {
+        $sqlcommand ="INSERT INTO evaluations (speaker, word, listener, evaluation, pronunc_variant, timestamp) ";
+        $sqlcommand.="VALUES ('$speaker', '$word', '$listener', '0', '$pronunc_variant', '$timestamp'); ";
+
+    }
+
+    #print $sqlcommand."<br>";
+
+
+    $trying=$db->exec($sqlcommand);
+
+    if ($trying)
+        print 1;
+    else
+        print 0;
+
+    $db->close();
+
+}
 
 ?>
